Directly add current coords to new discontinuous set
When using addCoords() on a continuous set of coordinates, the original
object will now be preserved as a subset of the new discontinuous set,
rather than being deleted/discarded as they originally were. This is
important for replacing and/or removing coordinates, since the method to
remove a set of coords works by reference; plus, just in theory, as much
of the structure should be preserved as is realistically possible from
operation to operation (that is, unnecessary changes to the objects
should be avoided).