Move header/closer to Slice, begin LyricInsertion class
Header and closer apply to all coordinate sets in a LyricSlice, so they
were moved to the LyricSlice class to line up with that idea
conceptually.

Changes to a reference string can now be tracked with a list of
LyricInsertion objects, which will store information about where to make
changes and what changes to make.